# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Code Reviewer API
  description: API Documentation for the AI Code Reviewer Platform. This API handles user authentication, project uploads (ZIP/GitHub), AI review trigger, and retrieval of review stats and chat.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local development server
paths:
...
  /projects:
    get:
      summary: Get all projects for the logged-in user
      tags:
        - Projects
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    projectId:
                      type: string
                      example: "123"
                    projectName:
                      type: string
                      example: Portfolio
                    status:
                      type: string
                      example: Completed
        '401':
          description: Unauthorized
...
  /chat:
    post:
      summary: Send a message to the AI assistant regarding a review
      tags:
        - AI Chat
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reviewId
                - message
              properties:
                reviewId:
                  type: string
                  example: "123"
                message:
                  type: string
                  example: Explain this function
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply:
                    type: string
                    example: This function validates JWT...
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT